Addition of Sym and some error correction
With the new "splitting" of addition, there are sometimes equations where the RHS is just a symbol. SCMC previously couldn't handle this properly; now it can. 
Errors included not referring to equation types correctly, and accidentally making convex/concave portions the same in all_evaluators, resulting in the convex functions not actually being convex. This has been fixed.